feat: move reference data, market data, dashboard, and other pages to feature modules
Relocates the following page directories into their corresponding feature modules:
- pages/reference-data/ → features/reference-data/pages/
- pages/market-data/ → features/market-data/pages/
- pages/forecasting/ → features/forecasting/pages/
- pages/internal-accounts/ → features/internal-accounts/pages/
- pages/audit/ → features/audit/pages/
- pages/dashboard/ → features/dashboard/pages/

Updates feature index.ts barrel exports and App.tsx imports to reference the new locations.
