+++ frontend/src/App.tsx
@@ -18,27 +18,22 @@ import { PaymentsPage } from '@/pages/payments'
 import { PaymentDetailPage } from '@/pages/payments/payment-detail'
 import { PartiesPage } from '@/pages/parties'
 import { PartyDetailPage } from '@/pages/parties/[partyId]'
-import { AuditLogPage } from '@/pages/audit'
+import { AuditLogPage } from '@/features/audit'
 import { StarlarkConfigPage } from '@/pages/starlark/index'
 import { StarlarkDetailPage } from '@/pages/starlark/detail'
 import { PositionsPage } from '@/pages/positions'
 import { PositionDetailPage } from '@/pages/positions/detail'
 import { MappingsPage } from '@/pages/mappings'
 import { MappingDetailPage } from '@/pages/mappings/[mappingId]'
-import { ReferenceDataHubPage } from '@/pages/reference-data'
-import { InstrumentsPage } from '@/pages/reference-data/instruments'
-import { AccountTypesPage } from '@/pages/reference-data/account-types'
-import { NodesPage } from '@/pages/reference-data/nodes'
-import { InternalAccountsPage } from '@/pages/internal-accounts'
-import { InternalAccountDetailPage } from '@/pages/internal-accounts/[accountId]'
-import { MarketDataPage } from '@/pages/market-data'
-import { DatasetDetailPage } from '@/pages/market-data/[datasetCode]'
-import { ForecastingPage } from '@/pages/forecasting'
+import { ReferenceDataHubPage, InstrumentsPage, AccountTypesPage, NodesPage } from '@/features/reference-data'
+import { InternalAccountsPage, InternalAccountDetailPage } from '@/features/internal-accounts'
+import { MarketDataPage, DatasetDetailPage } from '@/features/market-data'
+import { ForecastingPage } from '@/features/forecasting'
 import { LedgerPage } from '@/pages/ledger'
 import { BookingLogDetailPage } from '@/pages/ledger/booking-log-detail'
 import { ReconciliationPage } from '@/pages/reconciliation'
 import { ReconciliationDetailPage } from '@/pages/reconciliation/detail'
-import { DashboardPage } from '@/pages/dashboard'
+import { DashboardPage } from '@/features/dashboard'
 import { ManifestsPage } from '@/pages/manifests'
 import { McpConfigPage } from '@/pages/mcp-config'
 import { TransactionsPage } from '@/pages/transactions'
