+++ frontend/src/features/audit/index.ts
@@ -1,1 +1,2 @@
 // Public API for audit module
+export { AuditLogPage } from './pages/index'
