+++ frontend/src/features/dashboard/index.ts
@@ -1,1 +1,2 @@
 // Public API for dashboard module
+export { DashboardPage } from './pages/index'
