+++ frontend/src/features/dashboard/pages/quick-actions.tsx
