+++ frontend/src/features/dashboard/pages/stat-card.tsx
