+++ frontend/src/features/forecasting/index.ts
@@ -1,1 +1,2 @@
 // Public API for forecasting module
+export { ForecastingPage } from './pages/index'
