+++ frontend/src/features/forecasting/pages/index.tsx
