+++ frontend/src/features/internal-accounts/index.ts
@@ -1,1 +1,3 @@
 // Public API for internal-accounts module
+export { InternalAccountsPage } from './pages/index'
+export { InternalAccountDetailPage } from './pages/[accountId]'
