+++ frontend/src/features/internal-accounts/pages/[accountId].tsx
