+++ frontend/src/features/market-data/index.ts
@@ -1,1 +1,3 @@
 // Public API for market-data module
+export { MarketDataPage } from './pages/index'
+export { DatasetDetailPage } from './pages/[datasetCode]'
