+++ frontend/src/features/market-data/pages/constants.ts
