+++ frontend/src/features/reference-data/index.ts
@@ -1,1 +1,5 @@
 // Public API for reference-data module
+export { ReferenceDataHubPage } from './pages/index'
+export { InstrumentsPage } from './pages/instruments/index'
+export { AccountTypesPage } from './pages/account-types/index'
+export { NodesPage } from './pages/nodes/index'
