+++ frontend/src/features/reference-data/pages/instruments/index.tsx
