+++ frontend/src/features/reference-data/pages/instruments/register-instrument-dialog.test.tsx
