+++ frontend/src/features/reference-data/pages/nodes/create-node-dialog.tsx
